[Refactor] Update namespaces and tags for search controllers
