+++ src/Search/Controller/Asset/GetSearchResultController.php
@@ -14,18 +14,15 @@ declare(strict_types=1);
  *  @license    http://www.pimcore.org/license     GPLv3 and PCL
  */
 
-namespace Pimcore\Bundle\StudioBackendBundle\DataObject\Controller\Search;
+namespace Pimcore\Bundle\StudioBackendBundle\Search\Controller\Asset;
 
-use Exception;
 use OpenApi\Attributes\Post;
 use Pimcore\Bundle\StudioBackendBundle\Controller\AbstractApiController;
 use Pimcore\Bundle\StudioBackendBundle\Exception\Api\InvalidArgumentException;
-use Pimcore\Bundle\StudioBackendBundle\Exception\Api\NotFoundException;
 use Pimcore\Bundle\StudioBackendBundle\Grid\Attribute\Property\GridCollection;
 use Pimcore\Bundle\StudioBackendBundle\Grid\Attribute\Request\SearchGridRequestBody;
 use Pimcore\Bundle\StudioBackendBundle\Grid\MappedParameter\SearchGridParameter;
 use Pimcore\Bundle\StudioBackendBundle\Grid\Service\GridSearchServiceInterface;
-use Pimcore\Bundle\StudioBackendBundle\OpenApi\Attribute\Parameter\Path\StringParameter;
 use Pimcore\Bundle\StudioBackendBundle\OpenApi\Attribute\Response\Content\CollectionJson;
 use Pimcore\Bundle\StudioBackendBundle\OpenApi\Attribute\Response\DefaultResponses;
 use Pimcore\Bundle\StudioBackendBundle\OpenApi\Attribute\Response\SuccessResponse;
@@ -52,27 +49,19 @@ final class GetSearchResultController extends AbstractApiController
 
     /**
      * @throws InvalidArgumentException
-     * @throws Exception
-     * @throws NotFoundException
      */
-    #[Route('/data-objects/search/{classId?}', name: 'pimcore_studio_api_get_data_object_search', methods: ['POST'])]
-    #[IsGranted(UserPermissions::DATA_OBJECTS->value)]
+    #[Route('/assets/search', name: 'pimcore_studio_api_get_asset_search', methods: ['POST'])]
+    #[IsGranted(UserPermissions::ASSETS->value)]
     #[Post(
-        path: self::PREFIX . '/data-objects/search{classId}',
-        operationId: 'data_object_get_search',
-        description: 'data_object_get_search_description',
-        summary: 'data_object_get_search_summary',
-        tags: [Tags::DataObjectsSearch->value]
-    )]
-    #[StringParameter(
-        name: 'classId',
-        example: 'EV',
-        description: 'Class Id of the data object',
-        required: false,
+        path: self::PREFIX . '/assets/search',
+        operationId: 'asset_get_search',
+        description: 'asset_get_search_description',
+        summary: 'asset_get_search_summary',
+        tags: [Tags::Search->value]
     )]
     #[SearchGridRequestBody]
     #[SuccessResponse(
-        description: 'data_object_get_search_success_response',
+        description: 'asset_get_search_success_response',
         content: new CollectionJson(
             collection: new GridCollection()
         )
@@ -82,15 +71,8 @@ final class GetSearchResultController extends AbstractApiController
         HttpResponseCodes::NOT_FOUND,
         HttpResponseCodes::BAD_REQUEST,
     ])]
-    public function getDataObjectSearchGrid(
-        ?string $classId,
-        #[MapRequestPayload] SearchGridParameter $searchGridParameter
-    ): JsonResponse {
-        return $this->jsonResponse(
-            $this->searchService->getDataObjectSearchGrid(
-                $searchGridParameter,
-                $classId
-            )
-        );
+    public function getAssetSearchGrid(#[MapRequestPayload] SearchGridParameter $searchGridParameter): JsonResponse
+    {
+        return $this->jsonResponse($this->searchService->getAssetSearchGrid($searchGridParameter));
     }
 }
