+++ src/Search/Controller/DataObject/GetSearchResultController.php
@@ -14,15 +14,18 @@ declare(strict_types=1);
  *  @license    http://www.pimcore.org/license     GPLv3 and PCL
  */
 
-namespace Pimcore\Bundle\StudioBackendBundle\Asset\Controller\Search;
+namespace Pimcore\Bundle\StudioBackendBundle\Search\Controller\DataObject;
 
+use Exception;
 use OpenApi\Attributes\Post;
 use Pimcore\Bundle\StudioBackendBundle\Controller\AbstractApiController;
 use Pimcore\Bundle\StudioBackendBundle\Exception\Api\InvalidArgumentException;
+use Pimcore\Bundle\StudioBackendBundle\Exception\Api\NotFoundException;
 use Pimcore\Bundle\StudioBackendBundle\Grid\Attribute\Property\GridCollection;
 use Pimcore\Bundle\StudioBackendBundle\Grid\Attribute\Request\SearchGridRequestBody;
 use Pimcore\Bundle\StudioBackendBundle\Grid\MappedParameter\SearchGridParameter;
 use Pimcore\Bundle\StudioBackendBundle\Grid\Service\GridSearchServiceInterface;
+use Pimcore\Bundle\StudioBackendBundle\OpenApi\Attribute\Parameter\Path\StringParameter;
 use Pimcore\Bundle\StudioBackendBundle\OpenApi\Attribute\Response\Content\CollectionJson;
 use Pimcore\Bundle\StudioBackendBundle\OpenApi\Attribute\Response\DefaultResponses;
 use Pimcore\Bundle\StudioBackendBundle\OpenApi\Attribute\Response\SuccessResponse;
@@ -49,19 +52,27 @@ final class GetSearchResultController extends AbstractApiController
 
     /**
      * @throws InvalidArgumentException
+     * @throws Exception
+     * @throws NotFoundException
      */
-    #[Route('/assets/search', name: 'pimcore_studio_api_get_asset_search', methods: ['POST'])]
-    #[IsGranted(UserPermissions::ASSETS->value)]
+    #[Route('/data-objects/search/{classId?}', name: 'pimcore_studio_api_get_data_object_search', methods: ['POST'])]
+    #[IsGranted(UserPermissions::DATA_OBJECTS->value)]
     #[Post(
-        path: self::PREFIX . '/assets/search',
-        operationId: 'asset_get_search',
-        description: 'asset_get_search_description',
-        summary: 'asset_get_search_summary',
-        tags: [Tags::AssetSearch->value]
+        path: self::PREFIX . '/data-objects/search{classId}',
+        operationId: 'data_object_get_search',
+        description: 'data_object_get_search_description',
+        summary: 'data_object_get_search_summary',
+        tags: [Tags::Search->value]
+    )]
+    #[StringParameter(
+        name: 'classId',
+        example: 'EV',
+        description: 'Class Id of the data object',
+        required: false,
     )]
     #[SearchGridRequestBody]
     #[SuccessResponse(
-        description: 'asset_get_search_success_response',
+        description: 'data_object_get_search_success_response',
         content: new CollectionJson(
             collection: new GridCollection()
         )
@@ -71,8 +82,15 @@ final class GetSearchResultController extends AbstractApiController
         HttpResponseCodes::NOT_FOUND,
         HttpResponseCodes::BAD_REQUEST,
     ])]
-    public function getAssetSearchGrid(#[MapRequestPayload] SearchGridParameter $searchGridParameter): JsonResponse
-    {
-        return $this->jsonResponse($this->searchService->getAssetSearchGrid($searchGridParameter));
+    public function getDataObjectSearchGrid(
+        ?string $classId,
+        #[MapRequestPayload] SearchGridParameter $searchGridParameter
+    ): JsonResponse {
+        return $this->jsonResponse(
+            $this->searchService->getDataObjectSearchGrid(
+                $searchGridParameter,
+                $classId
+            )
+        );
     }
 }
